rename UI folder
