+++ plug-ins/video_command.py
@@ -1,18 +1,17 @@
 import maya.api.OpenMaya as om
 #from maya import OpenMayaMPx as mpx
 import maya.cmds as cmds
-import ui.main as player
+import video_command_ui.main as player
 
-# Not sure if this is necessary
+
 def maya_useNewAPI():
     pass
+
 
 def initializePlugin(plugin):
     vendor = "Derek Pepple"
     version = "1.0.0"
-
     pluging_fn = om.MFnPlugin(plugin, vendor, version)
-
     try:
         pluging_fn.registerCommand(VideoPlayerCmd.COMMAND_NAME, VideoPlayerCmd.creator)
     except:
@@ -21,7 +20,6 @@ def initializePlugin(plugin):
 
 def uninitializePlugin(plugin):
     plugin_fn = om.MFnPlugin(plugin)
-
     try:
         plugin_fn.deregisterCommand(VideoPlayerCmd.COMMAND_NAME)
     except:
@@ -34,10 +32,9 @@ class VideoPlayerCmd(om.MPxCommand):
 
     def __init__(self):
         super(VideoPlayerCmd, self).__init__()
-    
 
     def doIt(self, args):
-        print("doin it")
+        print("doing it")
         projectRootDirectory = cmds.workspace(q=True, rootDirectory=True)
         firstImagePath = cmds.renderSettings(firstImageName=True, fullPath=True)[0]
         path = cmds.renderSettings(genericFrameImageName="#",fullPath=True)[0]
