+++ plug-ins/video_command_ui/__pycache__/main.cpython-39.pyc
