+++ plug-ins/video_command_ui/__pycache__/resources_rc.cpython-39.pyc
