+++ plug-ins/video_command_ui/player/player/player.pyproject
