+++ plug-ins/video_command_ui/resources.qrc
