+++ plug-ins/video_command_ui/resources_rc.py
